Allows for input via stdin
main.cpp:
Reworks error messages
Changes case handling to allow for input on standard in

LZW-b.h:
Removes 'using namespace std' to avoid clobbering
Changes map to use uint16 instead of unsigned int

LZW-b.cpp
Changes shortToOutStream to potentially allow for printing to a given
file handle (currently uses stdout)
Adds ability to take input from stdin

+++ LZW-b.h
@@ -1,6 +1,6 @@
 #include <map>
 #include <string>
-//using namespace std;
+#include <stdint.h>
 
 class LZW{
   public:
@@ -16,6 +16,6 @@ class LZW{
     void extractLZW();
 
     std::map<std::string, unsigned int> codes;
-    std::map<unsigned int, std::string> strings;
+    std::map<uint16_t, std::string> strings;
 
 };
